updates to use lib comm

+++ REST Dashboard/Views/JoystickView.xaml.cs
@@ -28,13 +28,13 @@ namespace REST_Dashboard.Views
             SetY(127);
         }
 
-        public void SetX(float x)
+        public void SetX(double x)
         {
             int X_OFFSET = 40;
             Stick.Dispatcher.BeginInvoke((Action)(() => Canvas.SetLeft(Stick, X_OFFSET + ((x - 127)/127.0) * 50)));
         }
 
-        public void SetY(float y)
+        public void SetY(double y)
         {
             int Y_OFFSET = 40;
             Stick.Dispatcher.BeginInvoke((Action)(() => Canvas.SetTop(Stick, Y_OFFSET + ((y - 127) / 127.0) * 50)));
